+++ R/efficiency.R
@@ -401,3 +401,223 @@ calc_eta_fu_Y_eiou <- function(.c_mats_eta_phi_vecs = NULL,
 }
 
 
+#' Calculate aggregations and efficiencies
+#'
+#' Given a `matsindf`-style data frame
+#' of energy conversion chains (ECCs),
+#' this function calculates primary, final, useful, and services
+#' (when available) aggregates and associated efficiencies.
+#' Columns containing PSUT matrices are not returned in the output.
+#'
+#' Final, useful, and services data are assumed to be
+#' contained in the final demand matrix (**Y**) on various rows of `.psut_df`,
+#' identified by the `last_stage` column in `.psut_df`.
+#' This function will still work, even if primary energy
+#' is different for each last stage.
+#'
+#' Internally, primary aggregates are calculated
+#' using `primary_aggregates()`,
+#' final demand aggregates are calculated
+#' using `finaldemand_aggregates()`.
+#' The meaning of final demand aggregates for each row of `.psut_df`
+#' is determined by the corresponding value in the `last_stage` column.
+#'
+#' Note that when an ECC stage is not present,
+#' its aggregation and efficiency columns will be removed from output.
+#'
+#' If a services stage is present, its efficiencies will have mixed units
+#' and might be meaningless.
+#' Proceed with caution.
+#'
+#' @param .psut_df A data frame of energy conversion chain data in PSUT format.
+#' @param p_industries A string vector of primary industries.
+#' @param fd_sectors A string vector of final demand sectors.
+#' @param remove_psut_cols A boolean telling whether to delete columns containing
+#'                         PSUT matrices.
+#'                         Default is `TRUE`.
+#' @param piece The piece of the labels used for matching.
+#'              Default is "noun".
+#' @param notation The notation used for row and column labels.
+#'                 Default is `list(RCLabels::bracket_notation, RCLabels::arrow_notation)`.
+#' @param pattern_type The pattern type to be used for row and column matching.
+#'                     Default is "exact".
+#' @param prepositions A list of prepositions for row and column labels.
+#'                     Default is `RCLabels::prepositions_list`.
+#' @param R,U,V,Y,r_eiou,U_eiou,U_feed,S_units,last_stage String names of matrix columns in `.psut_df`.
+#'                                                        See `Recca::psut_cols`.
+#' @param gross,net,gross_net See `Recca::efficiency_cols`.
+#' @param primary,final,useful,services See `IEATools::all_stages`.
+#' @param ex_p,ex_fd_gross,ex_fd_net,ex_fd Names of aggregate columns. See `Recca::aggregate_cols`.
+#' @param ex_f,ex_u,ex_s See `IEATools::aggregate_cols`.
+#' @param eta_pf,eta_fu,eta_us,eta_pu,eta_ps,eta_fs See `Recca::efficiency_cols`.
+#'
+#' @return A data frame of metadata columns;
+#'         primary, final, useful, and services aggregations;
+#'         and efficiencies.
+#'
+#' @export
+#'
+#' @examples
+#' p_industries <- "Resources"
+#' fd_sectors <- c("Residential", "Transport", "Oil fields")
+#' UKEnergy2000mats |>
+#'   tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) |>
+#'   calc_agg_eta_pfus(p_industries = p_industries, fd_sectors = fd_sectors)
+calc_agg_eta_pfus <- function(.psut_df,
+                              p_industries,
+                              fd_sectors,
+                              remove_psut_cols = TRUE,
+                              piece = "noun",
+                              notation = list(RCLabels::bracket_notation,
+                                              RCLabels::arrow_notation),
+                              pattern_type = "exact",
+                              prepositions = RCLabels::prepositions_list,
+                              # Names of original matrices in .psut_data
+                              R = Recca::psut_cols$R,
+                              U = Recca::psut_cols$U,
+                              U_feed = Recca::psut_cols$U_feed,
+                              U_eiou = Recca::psut_cols$U_eiou,
+                              r_eiou = Recca::psut_cols$r_eiou,
+                              V = Recca::psut_cols$V,
+                              Y = Recca::psut_cols$Y,
+                              S_units = Recca::psut_cols$S_units,
+                              # Key names
+                              gross = Recca::efficiency_cols$gross,
+                              net = Recca::efficiency_cols$net,
+                              gross_net = Recca::efficiency_cols$gross_net,
+                              last_stage = Recca::psut_cols$last_stage,
+                              primary = Recca::all_stages$primary,
+                              final = Recca::all_stages$final,
+                              useful = Recca::all_stages$useful,
+                              services = Recca::all_stages$services,
+                              ex_p = Recca::aggregate_cols$aggregate_primary,
+                              ex_f = Recca::aggregate_cols$aggregate_final,
+                              ex_u = Recca::aggregate_cols$aggregate_useful,
+                              ex_s = Recca::aggregate_cols$aggregate_services,
+                              ex_fd_gross = Recca::aggregate_cols$gross_aggregate_demand,
+                              ex_fd_net = Recca::aggregate_cols$net_aggregate_demand,
+                              ex_fd = Recca::aggregate_cols$aggregate_demand,
+                              eta_pf = Recca::efficiency_cols$eta_pf,
+                              eta_fu = Recca::efficiency_cols$eta_fu,
+                              eta_us = Recca::efficiency_cols$eta_us,
+                              eta_pu = Recca::efficiency_cols$eta_pu,
+                              eta_ps = Recca::efficiency_cols$eta_ps,
+                              eta_fs = Recca::efficiency_cols$eta_fs) {
+
+  # Calculate primary aggregates
+  p_aggs <- .psut_df |>
+    Recca::primary_aggregates(p_industries = p_industries,
+                              piece = piece,
+                              notation = notation,
+                              pattern_type = pattern_type,
+                              prepositions = prepositions)
+
+  # Add final demand aggregates to the data frame
+  pfd_aggs <- p_aggs |>
+    Recca::finaldemand_aggregates(fd_sectors = fd_sectors,
+                                  piece = piece,
+                                  notation = notation,
+                                  pattern_type = pattern_type,
+                                  prepositions = prepositions) |>
+    dplyr::mutate(
+      "{R}" := NULL,
+      "{U}" := NULL,
+      "{U_feed}" := NULL,
+      "{U_eiou}" := NULL,
+      "{r_eiou}" := NULL,
+      "{V}" := NULL,
+      "{Y}" := NULL,
+      "{S_units}" := NULL
+    )
+
+  # Create a data frame of gross and net versions of
+  # primary and final demand energy.
+  gross_net_p_fd <- pfd_aggs |>
+    # Pivot to gross and net primary and final demand energy stages
+    dplyr::rename(
+      "{gross}" := dplyr::all_of(ex_fd_gross),
+      "{net}" := dplyr::all_of(ex_fd_net)
+    ) |>
+    tidyr::pivot_longer(cols = dplyr::any_of(c(gross, net)), names_to = gross_net, values_to = ex_fd)
+
+  # Isolate only the final demand energy to work on final and useful stages.
+  gross_net_fus <- gross_net_p_fd |>
+    dplyr::mutate(
+      "{ex_p}" := NULL
+    ) |>
+    tidyr::pivot_wider(names_from = dplyr::all_of(last_stage), values_from = dplyr::all_of(ex_fd)) |>
+    dplyr::rename(
+      "{ex_f}" := dplyr::any_of(final),
+      "{ex_u}" := dplyr::any_of(useful),
+      "{ex_s}" := dplyr::any_of(services)
+    )
+
+  # Isolate only primary aggregatges.
+  gross_net_p <- gross_net_p_fd |>
+    dplyr::mutate(
+      "{ex_fd}" := NULL
+    )
+
+  # Define a helpful selector function for later.
+  # This function will be used to select columns where
+  # not all of the entries are NA.
+  # I.e., it will delete columns where all entries are NA.
+  not_all_na_func <- function(x) {
+    any(!is.na(x))
+  }
+
+  # Join primary and final/useful/services, calculate efficiencies, and return,
+  # being careful to preserve all metadata columns.
+  out <- dplyr::full_join(gross_net_p, gross_net_fus, by = names(gross_net_p) |> setdiff(c(ex_p, last_stage)))
+  # Get the column names to do conditionals below.
+  cnames <- names(out)
+
+  # Calculate efficiencies,
+  # being careful to only calculate those with meaning.
+  if (all(c(ex_p, ex_f) %in% cnames)) {
+    out <- out |>
+      dplyr::mutate(
+        "{eta_pf}" := .data[[ex_f]] / .data[[ex_p]]
+      )
+  }
+  if (all(c(ex_f, ex_u) %in% cnames)) {
+    out <- out |>
+      dplyr::mutate(
+        "{eta_fu}" := .data[[ex_u]] / .data[[ex_f]]
+      )
+  }
+  if (all(c(ex_p, ex_u) %in% cnames)) {
+    out <- out |>
+      dplyr::mutate(
+        "{eta_pu}" := .data[[ex_u]] / .data[[ex_p]]
+      )
+  }
+  if (all(c(ex_p, ex_s) %in% cnames)) {
+    out <- out |>
+      dplyr::mutate(
+        "{eta_ps}" := .data[[ex_s]] / .data[[ex_p]]
+      )
+  }
+  if (all(c(ex_f, ex_s) %in% cnames)) {
+    out <- out |>
+      dplyr::mutate(
+        "{eta_fs}" := .data[[ex_s]] / .data[[ex_f]]
+      )
+  }
+  if (all(c(ex_s, ex_u) %in% cnames)) {
+    out <- out |>
+      dplyr::mutate(
+        "{eta_us}" := .data[[ex_s]] / .data[[ex_u]]
+      )
+  }
+  # Reorder columns,
+  # using any_of() to avoid errors when columns don't exist.
+  out |>
+    dplyr::select(-dplyr::any_of(c(ex_p, ex_f, ex_u, ex_s,
+                                   eta_pf, eta_fu, eta_pu,
+                                   eta_ps, eta_fs, eta_us)),
+                  dplyr::everything(),
+                  dplyr::any_of(c(ex_p, ex_f, ex_u, ex_s,
+                                  eta_pf, eta_fu, eta_pu,
+                                  eta_ps, eta_fs, eta_us)))
+}