+++ NEWS.md
@@ -8,6 +8,28 @@ Cite all releases with doi [10.5281/zenodo.5226085](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+
+# Recca 0.1.51 (2023-11-15) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10137045.svg)](https://doi.org/10.5281/zenodo.10137045)
+
+* Fixed a bug in `calc_agg_eta_pfus()` where
+  efficiencies were calculated incorrectly when 
+  primary energy was different
+  depending on the value of the `Last.stage` column.
+* No new tests.
+    * Still at 748 tests, all passing.
+    * Test coverage remains at 100%.
+
+
+# Recca 0.1.50 (2023-11-15) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10136470.svg)](https://doi.org/10.5281/zenodo.10136470)
+
+* New function `calc_agg_eta_pfus()` calculates
+  primary, final, useful, and services aggregates and
+  associated efficiencies.
+* New tests for new features.
+    * Now up to 748 tests, all passing.
+    * Test coverage remains at 100%.
+
+
 # Recca 0.1.49 (2023-11-03) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10070427.svg)](https://doi.org/10.5281/zenodo.10070427)
 
 * `remove_neu()` now also removes non-energy use products in the rows of **Y**.